backend/agent/requirements.txt
agno>=2.5.0
openai
sqlalchemy
mcp
fastapi
uvicorn[standard]
httpx
psycopg[binary]
python-dotenv
pyjwt>=2.8.0
bcrypt>=4.0.0
